Remove PHP_VERSION_ID checks
