+++ phpstan.neon.dist
@@ -73,10 +73,7 @@ parameters:
 		-
 			message: "#Call to function method_exists\\(\\) with ApiPlatform\\\\JsonApi\\\\Serializer\\\\ItemNormalizer and 'setCircularReferenc…' will always evaluate to false\\.#"
 			path: tests/JsonApi/Serializer/ItemNormalizerTest.php
-		# TODO: remove in v3.0
-		-
-			message: "#Dead catch - JsonException is never thrown in the try block.#"
-			path: src/Symfony/Bundle/Test/Response.php
+
 		# See https://github.com/phpstan/phpstan-symfony/issues/27
 		-
 			message: '#^Service "[^"]+" is private.$#'
