+++ src/Metadata/Resource/Factory/AttributesResourceNameCollectionFactory.php
@@ -52,7 +52,7 @@ final class AttributesResourceNameCollectionFactory implements ResourceNameColle
         }
 
         foreach (ReflectionClassRecursiveIterator::getReflectionClassesFromDirectories($this->paths) as $className => $reflectionClass) {
-            if (\PHP_VERSION_ID >= 80000 && $this->isResource($reflectionClass)) {
+            if ($this->isResource($reflectionClass)) {
                 $classes[$className] = true;
             }
         }
