+++ tests/Fixtures/TestBundle/Entity/SymfonyUuidDummy.php
@@ -15,45 +15,41 @@ namespace ApiPlatform\Tests\Fixtures\TestBundle\Entity;
 
 use ApiPlatform\Metadata\ApiResource;
 use Doctrine\ORM\Mapping as ORM;
-use Symfony\Bridge\Doctrine\Types\UuidType;
 use Symfony\Component\Uid\Uuid;
 
-/* @TODO remove this check in 3.0 */
-if (\PHP_VERSION_ID >= 70200 && class_exists(Uuid::class) && class_exists(UuidType::class)) {
-    /**
-     * @author Vincent Chalamon <vincentchalamon@gmail.com>
-     */
-    #[ApiResource]
-    #[ORM\Entity]
-    class SymfonyUuidDummy
+/**
+ * @author Vincent Chalamon <vincentchalamon@gmail.com>
+ */
+#[ApiResource]
+#[ORM\Entity]
+class SymfonyUuidDummy
+{
+    #[ORM\Id]
+    #[ORM\Column(type: 'symfony_uuid', unique: true)]
+    #[ORM\GeneratedValue(strategy: 'NONE')]
+    private $id;
+    #[ORM\Column(nullable: true)]
+    private $number;
+
+    public function __construct(?Uuid $id = null)
     {
-        #[ORM\Id]
-        #[ORM\Column(type: 'symfony_uuid', unique: true)]
-        #[ORM\GeneratedValue(strategy: 'NONE')]
-        private $id;
-        #[ORM\Column(nullable: true)]
-        private $number;
+        $this->id = $id ?? Uuid::v4();
+    }
 
-        public function __construct(?Uuid $id = null)
-        {
-            $this->id = $id ?? Uuid::v4();
-        }
+    public function getId(): Uuid
+    {
+        return $this->id;
+    }
 
-        public function getId(): Uuid
-        {
-            return $this->id;
-        }
+    public function getNumber(): ?string
+    {
+        return $this->number;
+    }
 
-        public function getNumber(): ?string
-        {
-            return $this->number;
-        }
+    public function setNumber(?string $number): self
+    {
+        $this->number = $number;
 
-        public function setNumber(?string $number): self
-        {
-            $this->number = $number;
-
-            return $this;
-        }
+        return $this;
     }
 }
